feat: simplify next auth config
